+++ lib/easy_theme.dart
@@ -1,3 +1,3 @@
-library easy_theme;
+library easy_design;
 
 export 'src/typography_screen.dart';